Moved test to more appropriate class.
